+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/cxxKinetics.cxx
@@ -524,7 +524,7 @@ Current_step(bool incremental_reactions, int reaction_step) const
 			}
 			else
 			{
-				kin_time = this->steps[reaction_step - 1];
+				kin_time = this->steps[(size_t)reaction_step - 1];
 			}
 		}
 		else 
@@ -550,7 +550,7 @@ Current_step(bool incremental_reactions, int reaction_step) const
 			}
 			else
 			{
-				kin_time = this->steps[reaction_step - 1];
+				kin_time = this->steps[(size_t)reaction_step - (size_t)1];
 			}
 		}
 		else
